remove function and implement state management instead
removes the function that checks whether a user has already requested a friendship with another user, and instead
uses a state variable so that the page updates when the data is retrieved. This variable is passed as a prop to the friendship
button component and determines whether it is disabled, or allows for a user to send a friend request.