update refresh process when user changes profile picture
Previously the page was refreshed, making new calls to the api. That was a hacky approach. In order to facilitate a more graceful process of changing a profile picture, the state that held the 'needsUpdate' status
was lifted up to the App component, and passed down to all the child components that consume that state. The Account Settings component
previously made a separate call to the api. This was removed completely, and the state data that it consumes (user name & profile picture) passed down to it from the App.

These changes help to minimize calls to the api, and prevent the page from being refreshed unnecessarily.