fix bug that broke layout
adds class name to wrapper div to fix layout-breaking change added with addition of infinite scroll feature.
When infinite scroll was added, what was the content wrapper had a grid-area property defined, and this was 'cancelled out' by the addition of the infinite scroll div elements. This
fix adds a wrapper around the entire component, and adds a className prop which re-adds the grid-area property and fixes the broken layout.